Removed movieId from movies.cs, initialized database and auto-added movies in program.cs, and added methods in databasehelper.cs for reading JSON files and executing queries
Removed movieId from movies.cs as it is automatically generated in the database. In program.cs, initialize the database and automatically add the movies. In databasehelper.cs, created methods to read JSON files and add them to the database, also created an ExecuteQuery method to retrieve information from the database.